Move ActionType functionality into the action class.
I remembered that action types are not a really useful concept for now
they are just wrappers for the plugin metadata.

This also saves us some major headaches for migrating the old email to
target content-types (they would be missing the action type class in
hook_campaignion_action_info().
